refactor(openclaw): reduce settings complexity
