+++ gui/src/main/handlers/settings.ts
@@ -8,16 +8,11 @@ import { buildAiSettings } from '../aiSettings'
 import { buildVoiceSettings, buildWakeWordStatus } from '../voiceSettings'
 import { defaultSettingsMap } from '../settingsDefaults'
 import { mirrorToRexConfig } from '../settingsMirror'
-import { reconcileIntegrationStatuses } from '../integrationStatus'
 import { vaultDeleteSecret, vaultHasSecret, vaultSetSecret, type VaultContext } from '../credentialVault'
-import {
-  deleteVaultReference,
-  getVaultReference,
-  putVaultReference,
-  type VaultReferenceRecord
-} from '../credentialReferences'
+import { getVaultReference, putVaultReference } from '../credentialReferences'
 import type { ElectronSessionIdentity } from '../sessionIdentity'
 import { safeIpcErrorMessage, SafeValidationError } from '../ipcErrors'
+import { loadIntegrationSettings, persistSettingsSection, removeEmailAccount } from '../integrationSettingsStorage'
 
 const ALLOWED_API_KEYS = [
   'OPENAI_API_KEY',
@@ -34,27 +29,6 @@ const ALLOWED_API_KEYS = [
   'OPENCLAW_GATEWAY_TOKEN'
 ] // pragma: allowlist secret
 
-interface SecretFieldSpec {
-  field: string
-  logicalName: string
-  context: VaultContext
-}
-
-const INTEGRATION_SECRET_FIELDS: SecretFieldSpec[] = [
-  { field: 'emailClientSecret', logicalName: 'EMAIL_CLIENT_SECRET', context: { scope: 'user', integration: 'email', account: 'default', slot: 'client_secret' } },
-  { field: 'calendarClientSecret', logicalName: 'CALENDAR_CLIENT_SECRET', context: { scope: 'user', integration: 'calendar', account: 'default', slot: 'client_secret' } },
-  { field: 'smsSid', logicalName: 'TWILIO_ACCOUNT_SID', context: { scope: 'household', integration: 'twilio', account: 'sms', slot: 'account_sid' } },
-  { field: 'smsAuthToken', logicalName: 'TWILIO_AUTH_TOKEN', context: { scope: 'household', integration: 'twilio', account: 'sms', slot: 'auth_token' } },
-  { field: 'smsFromNumber', logicalName: 'TWILIO_FROM_NUMBER', context: { scope: 'household', integration: 'twilio', account: 'sms', slot: 'from_number' } },
-  { field: 'haToken', logicalName: 'HA_TOKEN', context: { scope: 'household', integration: 'home_assistant', account: null, slot: 'token' } },
-  { field: 'phoneSid', logicalName: 'TWILIO_PHONE_ACCOUNT_SID', context: { scope: 'household', integration: 'twilio', account: 'phone', slot: 'account_sid' } },
-  { field: 'phoneAuthToken', logicalName: 'TWILIO_PHONE_AUTH_TOKEN', context: { scope: 'household', integration: 'twilio', account: 'phone', slot: 'auth_token' } },
-  { field: 'phoneNumber', logicalName: 'TWILIO_PHONE_NUMBER', context: { scope: 'household', integration: 'twilio', account: 'phone', slot: 'phone_number' } },
-  { field: 'phoneTransferNumber', logicalName: 'TWILIO_TRANSFER_NUMBER', context: { scope: 'household', integration: 'twilio', account: 'phone', slot: 'transfer_number' } },
-  { field: 'telegramBotToken', logicalName: 'TELEGRAM_BOT_TOKEN', context: { scope: 'household', integration: 'telegram', account: null, slot: 'token' } },
-  { field: 'openclawToken', logicalName: 'OPENCLAW_GATEWAY_TOKEN', context: { scope: 'household', integration: 'openclaw_gateway', account: null, slot: 'token' } }
-]
-
 /** Derive a short integration name for vault metadata, e.g. OPENAI_API_KEY -> openai. */
 function integrationNameForKey(key: string): string {
   return key.replace(/_API_KEY$|_KEY$/, '').toLowerCase()
@@ -67,14 +41,6 @@ function apiKeyContext(key: string): VaultContext {
   return { scope: 'household', integration: integrationNameForKey(key), account: null, slot: 'api_key' }
 }
 
-function validateAccountId(value: unknown): string {
-  const id = typeof value === 'string' ? value : ''
-  if (!/^[A-Za-z0-9][A-Za-z0-9._:@-]{0,127}$/.test(id)) {
-    throw new Error('Email account ID is invalid')
-  }
-  return id
-}
-
 export function registerSettingsHandlers(session: ElectronSessionIdentity): void {
   ipcMain.handle('rex:getSettings', async (_event, section: string): Promise<Settings> => {
     const stored = readGuiSettings()
@@ -84,60 +50,10 @@ export function registerSettingsHandlers(session: ElectronSessionIdentity): void
     if (section === 'voice') {
       return buildVoiceSettings((stored[section] ?? {}) as Settings) as unknown as Settings
     }
-    if (section !== 'integrations') return stored[section] ?? defaultSettingsMap[section] ?? {}
-
-    const integrations = {
-      ...defaultSettingsMap.integrations,
-      ...((stored.integrations ?? {}) as Record<string, unknown>)
-    } as Record<string, unknown>
-    const config = readRexConfigStrict()
-    const explicitIntegrations = (stored.integrations ?? {}) as Record<string, unknown>
-    const openclaw = config.openclaw && typeof config.openclaw === 'object'
-      ? config.openclaw as Record<string, unknown>
-      : {}
-    if (!Object.prototype.hasOwnProperty.call(explicitIntegrations, 'openclawGatewayUrl')) {
-      integrations.openclawGatewayUrl =
-        typeof openclaw.gateway_url === 'string' ? openclaw.gateway_url : ''
+    if (section === 'integrations') {
+      return loadIntegrationSettings(session, stored as Record<string, Settings>)
     }
-    if (!Object.prototype.hasOwnProperty.call(explicitIntegrations, 'openclawToolsEnabled')) {
-      integrations.openclawToolsEnabled = openclaw.use_tools === true
-    }
-    if (!Object.prototype.hasOwnProperty.call(explicitIntegrations, 'openclawVoiceEnabled')) {
-      integrations.openclawVoiceEnabled = openclaw.use_voice_backend === true
-    }
-    const credentialStatus: Record<string, { ref: string; hasCredential: boolean }> = {}
-    for (const spec of INTEGRATION_SECRET_FIELDS) {
-      integrations[spec.field] = ''
-      const record = getVaultReference(config, spec.logicalName, spec.context, session.userId)
-      if (record) {
-        credentialStatus[spec.field] = {
-          ref: record.ref,
-          hasCredential: await vaultHasSecret(session, record.ref, spec.context)
-        }
-      }
-    }
-    const accounts = Array.isArray(integrations.emailAccounts) ? integrations.emailAccounts : []
-    integrations.emailAccounts = await Promise.all(accounts.map(async (raw) => {
-      const account: Record<string, unknown> = {
-        ...(raw as Record<string, unknown>),
-        password: '',
-        clientSecret: ''
-      }
-      const id = validateAccountId(account.id)
-      const backend = account.backend === 'imap' ? 'imap' : account.backend === 'outlook' ? 'outlook' : 'gmail'
-      const logicalName = `email:${id}`
-      const context: VaultContext = { scope: 'user', integration: 'email', account: id, slot: backend === 'imap' ? 'password' : 'client_secret' }
-      const record = getVaultReference(config, logicalName, context, session.userId)
-      if (record) {
-        account.credentialRef = record.ref
-        account.hasCredential = await vaultHasSecret(session, record.ref, context)
-      } else {
-        account.hasCredential = false
-      }
-      return account
-    }))
-    integrations.credentialStatus = credentialStatus
-    return integrations as Settings
+    return stored[section] ?? defaultSettingsMap[section] ?? {}
   })
 
   ipcMain.handle('rex:getWakeWordStatus', (_event, values?: Settings): WakeWordStatus => {
@@ -149,134 +65,13 @@ export function registerSettingsHandlers(session: ElectronSessionIdentity): void
   ipcMain.handle(
     'rex:setSettings',
     async (_event, section: string, values: Settings): Promise<{ ok: boolean; error?: string }> => {
-      const stored = readGuiSettings()
-      const originalStored = JSON.parse(JSON.stringify(stored)) as Record<string, Settings>
-      const originalConfig = readRexConfigStrict()
-      const nextConfig = JSON.parse(JSON.stringify(originalConfig)) as Record<string, unknown>
-      let normalizedValues =
+      const normalizedValues =
         section === 'ai'
           ? (buildAiSettings(values) as unknown as Settings)
           : section === 'voice'
             ? (buildVoiceSettings(values) as unknown as Settings)
             : values
-      const newEntries: Array<{ logicalName: string; ref: string; context: VaultContext }> = []
-      const replacedEntries: Array<{ record: VaultReferenceRecord; context: VaultContext }> = []
-      let guiWritten = false
-      let configWritten = false
-      try {
-        if (section === 'integrations') {
-          const raw = { ...(normalizedValues as unknown as Record<string, unknown>) }
-          delete raw.credentialStatus
-          for (const spec of INTEGRATION_SECRET_FIELDS) {
-            const oldRecord = getVaultReference(
-              originalConfig, spec.logicalName, spec.context, session.userId
-            )
-            const secret = typeof raw[spec.field] === 'string' ? (raw[spec.field] as string) : ''
-            if (secret.trim()) {
-              const ref = await vaultSetSecret(session, secret, spec.context)
-              newEntries.push({ logicalName: spec.logicalName, ref, context: spec.context })
-              putVaultReference(nextConfig, spec.logicalName, ref, spec.context, session.userId)
-              if (oldRecord) replacedEntries.push({ record: oldRecord, context: spec.context })
-            }
-            raw[spec.field] = ''
-          }
-
-          const accounts = Array.isArray(raw.emailAccounts) ? raw.emailAccounts : []
-          const originalIntegrations = (originalStored.integrations ?? {}) as Record<string, unknown>
-          const originalAccounts = Array.isArray(originalIntegrations.emailAccounts)
-            ? originalIntegrations.emailAccounts
-            : []
-          const originalById = new Map<string, Record<string, unknown>>()
-          for (const candidate of originalAccounts) {
-            if (!candidate || typeof candidate !== 'object') continue
-            const originalAccount = candidate as Record<string, unknown>
-            originalById.set(validateAccountId(originalAccount.id), originalAccount)
-          }
-          raw.emailAccounts = []
-          const seenAccountIds = new Set<string>()
-          for (const rawAccount of accounts) {
-            if (!rawAccount || typeof rawAccount !== 'object') continue
-            const id = validateAccountId((rawAccount as Record<string, unknown>).id)
-            if (seenAccountIds.has(id)) throw new SafeValidationError(`Duplicate email account id: ${id}`)
-            seenAccountIds.add(id)
-          }
-          for (const rawAccount of accounts) {
-            if (!rawAccount || typeof rawAccount !== 'object') continue
-            const account = { ...(rawAccount as Record<string, unknown>) }
-            const id = validateAccountId(account.id)
-            const backend = account.backend === 'imap' ? 'imap' : account.backend === 'outlook' ? 'outlook' : 'gmail'
-            const context: VaultContext = {
-              scope: 'user', integration: 'email', account: id,
-              slot: backend === 'imap' ? 'password' : 'client_secret'
-            }
-            const logicalName = `email:${id}`
-            const originalAccount = originalById.get(id)
-            const oldBackend = originalAccount?.backend === 'imap'
-              ? 'imap'
-              : originalAccount?.backend === 'outlook' ? 'outlook' : 'gmail'
-            const oldContext: VaultContext = {
-              scope: 'user', integration: 'email', account: id,
-              slot: oldBackend === 'imap' ? 'password' : 'client_secret'
-            }
-            const oldRecord = getVaultReference(
-              originalConfig, logicalName, oldContext, session.userId
-            )
-            const field = backend === 'imap' ? 'password' : 'clientSecret'
-            const secret = typeof account[field] === 'string' ? (account[field] as string) : ''
-            if (oldRecord && oldContext.slot !== context.slot && !secret.trim()) {
-              throw new SafeValidationError(
-                'A new credential is required when changing an email account backend'
-              )
-            }
-            if (secret.trim()) {
-              const ref = await vaultSetSecret(session, secret, context)
-              newEntries.push({ logicalName, ref, context })
-              putVaultReference(nextConfig, logicalName, ref, context, session.userId)
-              if (oldRecord) replacedEntries.push({ record: oldRecord, context: oldContext })
-            }
-            delete account.password
-            delete account.clientSecret
-            delete account.credentialRef
-            delete account.hasCredential
-            ;(raw.emailAccounts as Record<string, unknown>[]).push(account)
-          }
-          normalizedValues = raw as Settings
-        }
-
-        stored[section] = normalizedValues
-        writeGuiSettings(stored)
-        guiWritten = true
-        writeRexConfig(nextConfig)
-        configWritten = true
-        const mirrorResult = mirrorToRexConfig(section, normalizedValues)
-        if (!mirrorResult.ok) throw new Error(mirrorResult.error ?? 'Settings mirror failed')
-        const readback = readRexConfigStrict()
-        for (const entry of newEntries) {
-          const record = getVaultReference(
-            readback, entry.logicalName, entry.context, session.userId
-          )
-          if (record?.ref !== entry.ref) throw new SafeValidationError('Credential reference readback failed')
-        }
-        if (section === 'integrations') await reconcileIntegrationStatuses(session)
-        for (const old of replacedEntries) {
-          await vaultDeleteSecret(session, old.record.ref, old.context).catch(() => false)
-        }
-        return { ok: true }
-      } catch (err) {
-        let restored = true
-        if (guiWritten) {
-          try { writeGuiSettings(originalStored) } catch { restored = false }
-        }
-        if (configWritten) {
-          try { writeRexConfig(originalConfig) } catch { restored = false }
-        }
-        if (restored) {
-          for (const entry of newEntries) {
-            await vaultDeleteSecret(session, entry.ref, entry.context).catch(() => false)
-          }
-        }
-        return { ok: false, error: safeIpcErrorMessage(err, 'Settings persistence failed') }
-      }
+      return persistSettingsSection(session, section, normalizedValues)
     }
   )
 
@@ -286,71 +81,7 @@ export function registerSettingsHandlers(session: ElectronSessionIdentity): void
       _event,
       idValue: string,
       confirmed: boolean
-    ): Promise<{ ok: boolean; error?: string }> => {
-      if (confirmed !== true) {
-        return { ok: false, error: 'Email account removal requires confirmation' }
-      }
-      let id: string
-      try {
-        id = validateAccountId(idValue)
-      } catch (err) {
-        return { ok: false, error: safeIpcErrorMessage(err, 'Email account ID is invalid') }
-      }
-
-      const originalStored = readGuiSettings()
-      const originalConfig = readRexConfigStrict()
-      const nextStored = JSON.parse(JSON.stringify(originalStored)) as Record<string, Settings>
-      const nextConfig = JSON.parse(JSON.stringify(originalConfig)) as Record<string, unknown>
-      const integrations = {
-        ...((nextStored.integrations ?? {}) as Record<string, unknown>)
-      }
-      const accounts = Array.isArray(integrations.emailAccounts)
-        ? integrations.emailAccounts as Array<Record<string, unknown>>
-        : []
-      const account = accounts.find((candidate) => candidate.id === id)
-      if (!account) return { ok: false, error: 'Email account not found' }
-      const backend = account.backend === 'imap'
-        ? 'imap'
-        : account.backend === 'outlook' ? 'outlook' : 'gmail'
-      const context: VaultContext = {
-        scope: 'user', integration: 'email', account: id,
-        slot: backend === 'imap' ? 'password' : 'client_secret'
-      }
-      const logicalName = `email:${id}`
-      let record: VaultReferenceRecord | null
-      try {
-        record = getVaultReference(originalConfig, logicalName, context, session.userId)
-      } catch (err) {
-        return { ok: false, error: safeIpcErrorMessage(err, 'Credential reference is invalid') }
-      }
-      integrations.emailAccounts = accounts.filter((candidate) => candidate.id !== id)
-      nextStored.integrations = integrations as Settings
-      if (record) deleteVaultReference(nextConfig, logicalName, context, session.userId)
-
-      let guiWritten = false
-      let configWritten = false
-      try {
-        writeGuiSettings(nextStored)
-        guiWritten = true
-        writeRexConfig(nextConfig)
-        configWritten = true
-        const mirror = mirrorToRexConfig('integrations', integrations as Settings)
-        if (!mirror.ok) throw new Error(mirror.error ?? 'Settings mirror failed')
-        if (getVaultReference(readRexConfigStrict(), logicalName, context, session.userId)) {
-          throw new SafeValidationError('Credential reference deletion readback failed')
-        }
-        if (record) await vaultDeleteSecret(session, record.ref, context)
-        return { ok: true }
-      } catch (err) {
-        if (guiWritten) {
-          try { writeGuiSettings(originalStored) } catch { /* preserve deletion error */ }
-        }
-        if (configWritten) {
-          try { writeRexConfig(originalConfig) } catch { /* preserve deletion error */ }
-        }
-        return { ok: false, error: safeIpcErrorMessage(err, 'Email account removal failed') }
-      }
-    }
+    ): Promise<{ ok: boolean; error?: string }> => removeEmailAccount(session, idValue, confirmed)
   )
 
   ipcMain.handle('rex:testVoice', () => {
