+++ gui/src/pages/SettingsPage.tsx
@@ -1,7 +1,8 @@
 import React, { useState, useEffect, useRef } from 'react'
-import { NavLink } from 'react-router-dom'
-import type { GeneralSettings, VoiceSettings, AiSettings, IntegrationsSettings, EmailAccount, NotificationsSettings, Settings, VersionInfo, PreferenceSuggestion, VoiceInfo, WakeWordInfo, WakeWordStatus, VoiceEnrollment, Memory, SmartSpeaker, SystemSettings, IntegrationInventoryItem } from '../types/ipc'
+import type { GeneralSettings, VoiceSettings, AiSettings, NotificationsSettings, Settings, VersionInfo, PreferenceSuggestion, VoiceInfo, WakeWordInfo, WakeWordStatus, VoiceEnrollment, Memory, SmartSpeaker, SystemSettings } from '../types/ipc'
 import { useToast } from '../components/ui/Toast'
+import { IntegrationsSettingsSection } from './settings/integrations/IntegrationsSettingsSection'
+import { PasswordInput } from './settings/shared'
 import { PageLoadingFallback } from '../components/ui/PageLoadingFallback'
 import { SkeletonLine } from '../components/ui/SkeletonLine'
 import { Tooltip } from '../components/ui/Tooltip'
@@ -2740,219 +2741,6 @@ function AiPanel(): React.ReactElement {
   )
 }
 
-type IntegrationSection = 'email' | 'calendar' | 'sms' | 'homeassistant' | 'phone' | 'openclaw'
-type TestStatus = 'idle' | 'testing' | 'configured' | 'ok' | 'error'
-
-function integrationStatusToTestStatus(state: IntegrationInventoryItem['state']): TestStatus {
-  if (['authenticated', 'read_only', 'write_capable', 'write_tested', 'verified'].includes(state)) return 'ok'
-  if (state === 'configured' || state === 'reachable') return 'configured'
-  if (state === 'degraded' || state === 'unavailable') return 'error'
-  return 'idle'
-}
-
-function integrationKeyToSection(key: string): IntegrationSection | null {
-  if (key === 'email' || key === 'calendar' || key === 'sms' || key === 'homeassistant' || key === 'phone' || key === 'openclaw') {
-    return key
-  }
-  return null
-}
-
-const INTEGRATION_SECRET_FIELDS = new Set<keyof IntegrationsSettings>([
-  'emailClientSecret', 'calendarClientSecret', 'smsSid', 'smsAuthToken',
-  'smsFromNumber', 'haToken', 'phoneSid', 'phoneAuthToken', 'phoneNumber',
-  'phoneTransferNumber', 'telegramBotToken', 'openclawToken'
-])
-
-function isIntegrationSecretField(field: keyof IntegrationsSettings): boolean {
-  return INTEGRATION_SECRET_FIELDS.has(field)
-}
-
-function hasStoredCredential(form: IntegrationsSettings, field: keyof IntegrationsSettings): boolean {
-  return form.credentialStatus[field]?.hasCredential === true
-}
-
-function sectionsForIntegrationField(field: keyof IntegrationsSettings): IntegrationSection[] {
-  if (
-    field === 'emailProvider' ||
-    field === 'emailClientId' ||
-    field === 'emailClientSecret' ||
-    field === 'emailAccounts'
-  ) {
-    return ['email']
-  }
-  if (field === 'calendarProvider' || field === 'calendarClientId' || field === 'calendarClientSecret') {
-    return ['calendar']
-  }
-  if (field === 'smsSid' || field === 'smsAuthToken' || field === 'smsFromNumber') {
-    return ['sms']
-  }
-  if (field === 'haUrl' || field === 'haToken') {
-    return ['homeassistant']
-  }
-  if (field === 'phoneSid' || field === 'phoneAuthToken' || field === 'phoneNumber') {
-    return ['phone']
-  }
-  if (
-    field === 'openclawGatewayUrl' ||
-    field === 'openclawToolsEnabled' ||
-    field === 'openclawVoiceEnabled' ||
-    field === 'openclawToken'
-  ) {
-    return ['openclaw']
-  }
-  return []
-}
-
-function ConnectionBadge({
-  status,
-  hasCredentials
-}: {
-  status: TestStatus
-  hasCredentials: boolean
-}): React.ReactElement {
-  if (status === 'ok') {
-    return (
-      <span className="inline-flex items-center gap-1 rounded-full bg-success/15 px-2 py-0.5 text-xs font-medium text-success">
-        <span className="h-1.5 w-1.5 rounded-full bg-success" />
-        Live test passed
-      </span>
-    )
-  }
-  if (status === 'error') {
-    return (
-      <span className="inline-flex items-center gap-1 rounded-full bg-danger/15 px-2 py-0.5 text-xs font-medium text-danger">
-        <span className="h-1.5 w-1.5 rounded-full bg-danger" />
-        Error
-      </span>
-    )
-  }
-  if (!hasCredentials) {
-    return (
-      <span className="inline-flex items-center gap-1 rounded-full bg-border px-2 py-0.5 text-xs font-medium text-text-secondary">
-        <span className="h-1.5 w-1.5 rounded-full bg-text-secondary" />
-        Not Configured
-      </span>
-    )
-  }
-  return (
-    <span className="inline-flex items-center gap-1 rounded-full bg-warning/15 px-2 py-0.5 text-xs font-medium text-warning">
-      <span className="h-1.5 w-1.5 rounded-full bg-warning" />
-      Configured only
-    </span>
-  )
-}
-
-function EyeIcon({ open }: { open: boolean }): React.ReactElement {
-  if (open) {
-    return (
-      <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-        <path d="M1 12s4-8 11-8 11 8 11 8-4 8-11 8-11-8-11-8z" />
-        <circle cx="12" cy="12" r="3" />
-      </svg>
-    )
-  }
-  return (
-    <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-      <path d="M17.94 17.94A10.07 10.07 0 0 1 12 20c-7 0-11-8-11-8a18.45 18.45 0 0 1 5.06-5.94" />
-      <path d="M9.9 4.24A9.12 9.12 0 0 1 12 4c7 0 11 8 11 8a18.5 18.5 0 0 1-2.16 3.19" />
-      <line x1="1" y1="1" x2="23" y2="23" />
-    </svg>
-  )
-}
-
-function PasswordInput({
-  id,
-  value,
-  placeholder,
-  onChange,
-  onBlur
-}: {
-  id: string
-  value: string
-  placeholder?: string
-  onChange: (v: string) => void
-  onBlur?: () => void
-}): React.ReactElement {
-  const [show, setShow] = useState(false)
-  return (
-    <div className="relative">
-      <input
-        id={id}
-        type={show ? 'text' : 'password'}
-        value={value}
-        placeholder={placeholder}
-        onChange={(e) => onChange(e.target.value)}
-        onBlur={onBlur}
-        className="w-full bg-surface-raised border border-border rounded-lg px-3 py-2 pr-10 text-sm text-text-primary placeholder:text-text-secondary focus:outline-none focus:ring-2 focus:ring-accent"
-      />
-      <button
-        type="button"
-        onClick={() => setShow((s) => !s)}
-        className="absolute right-2.5 top-1/2 -translate-y-1/2 text-text-secondary hover:text-text-primary transition-colors"
-        aria-label={show ? 'Hide' : 'Show'}
-      >
-        <EyeIcon open={show} />
-      </button>
-    </div>
-  )
-}
-
-function TestConnectionButton({
-  status,
-  onTest,
-  error
-}: {
-  status: TestStatus
-  onTest: () => void
-  error?: string
-}): React.ReactElement {
-  return (
-    <div className="mt-3">
-      <div className="flex items-center gap-3">
-        <button
-          onClick={onTest}
-          disabled={status === 'testing'}
-          className="flex items-center gap-2 bg-surface-raised hover:bg-border disabled:opacity-50 text-text-primary text-xs font-medium px-3 py-1.5 rounded-lg border border-border transition-colors focus:outline-none focus:ring-2 focus:ring-accent"
-        >
-          {status === 'testing' ? (
-            <>
-              <svg className="animate-spin h-3 w-3" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                <path d="M12 2v4M12 18v4M4.93 4.93l2.83 2.83M16.24 16.24l2.83 2.83M2 12h4M18 12h4M4.93 19.07l2.83-2.83M16.24 7.76l2.83-2.83" />
-              </svg>
-              Testing…
-            </>
-          ) : 'Check Status'}
-        </button>
-        {status === 'ok' && (
-          <span className="flex items-center gap-1 text-xs text-success">
-            <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3">
-              <polyline points="20 6 9 17 4 12" />
-            </svg>
-            Authenticated
-          </span>
-        )}
-        {status === 'error' && (
-          <span className="flex items-center gap-1 text-xs text-danger">
-            <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3">
-              <line x1="18" y1="6" x2="6" y2="18" /><line x1="6" y1="6" x2="18" y2="18" />
-            </svg>
-            Check failed
-          </span>
-        )}
-        {status === 'configured' && (
-          <span className="text-xs text-warning">Configured only</span>
-        )}
-        {status === 'idle' && (
-          <span className="text-xs text-text-secondary">Not tested</span>
-        )}
-      </div>
-      {(status === 'error' || status === 'configured') && error && (
-        <p className={`mt-1 text-xs ${status === 'error' ? 'text-danger' : 'text-warning'}`}>{error}</p>
-      )}
-    </div>
-  )
-}
-
 function WakeWordStatusPanel({ status }: { status: WakeWordStatus | null }): React.ReactElement | null {
   if (!status) return null
 
@@ -2998,1172 +2786,7 @@ function WakeWordStatusPanel({ status }: { status: WakeWordStatus | null }): Rea
 }
 
 function IntegrationsPanel(): React.ReactElement {
-  const addToast = useToast()
-  const [form, setForm] = useState<IntegrationsSettings>({
-    emailProvider: 'gmail',
-    emailClientId: '',
-    emailClientSecret: '',
-    emailAccounts: [],
-    calendarProvider: 'gmail',
-    calendarClientId: '',
-    calendarClientSecret: '',
-    smsSid: '',
-    smsAuthToken: '',
-    smsFromNumber: '',
-    haUrl: '',
-    haToken: '',
-    phoneSid: '',
-    phoneAuthToken: '',
-    phoneNumber: '',
-    phoneTransferNumber: '',
-    voicemailNotificationsEnabled: false,
-    contactsFilePath: '',
-    telegramBotToken: '',
-    telegramChatId: '',
-    openclawGatewayUrl: '',
-    openclawToolsEnabled: false,
-    openclawVoiceEnabled: false,
-    openclawToken: '',
-    credentialStatus: {}
-  })
-  const [loading, setLoading] = useState(true)
-  const [savedField, setSavedField] = useState<keyof IntegrationsSettings | null>(null)
-  const [testStatus, setTestStatus] = useState<Record<IntegrationSection, TestStatus>>({
-    email: 'idle',
-    calendar: 'idle',
-    sms: 'idle',
-    homeassistant: 'idle',
-    phone: 'idle',
-    openclaw: 'idle'
-  })
-  const [testErrors, setTestErrors] = useState<Partial<Record<IntegrationSection, string>>>({})
-  const savedTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-  const [accountTestStatus, setAccountTestStatus] = useState<Record<string, TestStatus>>({})
-  const accountTestTimers = useRef<Record<string, ReturnType<typeof setTimeout>>>({})
-
-  function handleTestEmailAccount(id: string): void {
-    setAccountTestStatus((s) => ({ ...s, [id]: 'testing' }))
-    window.rex
-      .testEmailAccount(id)
-      .then((res) => {
-        const nextStatus = res.state
-          ? integrationStatusToTestStatus(res.state)
-          : res.ok ? 'ok' : 'error'
-        setAccountTestStatus((s) => ({ ...s, [id]: nextStatus }))
-      })
-      .catch(() => {
-        setAccountTestStatus((s) => ({ ...s, [id]: 'error' }))
-      })
-      .finally(() => {
-        if (accountTestTimers.current[id]) clearTimeout(accountTestTimers.current[id])
-        accountTestTimers.current[id] = setTimeout(
-          () => setAccountTestStatus((s) => ({ ...s, [id]: 'idle' })),
-          5000
-        )
-      })
-  }
-
-  function applyIntegrationStatuses(inventory: IntegrationInventoryItem[]): void {
-    setTestStatus((current) => {
-      const next = { ...current }
-      for (const item of inventory) {
-        const section = integrationKeyToSection(item.key)
-        if (section) next[section] = integrationStatusToTestStatus(item.state)
-      }
-      return next
-    })
-    setTestErrors((current) => {
-      const next = { ...current }
-      for (const item of inventory) {
-        const section = integrationKeyToSection(item.key)
-        if (!section) continue
-        if ((item.state === 'degraded' || item.state === 'unavailable') && item.error) {
-          next[section] = item.error
-        } else if (item.state !== 'degraded' && item.state !== 'unavailable') {
-          delete next[section]
-        }
-      }
-      return next
-    })
-  }
-
-  function resetStatusForField(field: keyof IntegrationsSettings): void {
-    const sections = sectionsForIntegrationField(field)
-    if (sections.length === 0) return
-    setTestStatus((current) => {
-      const next = { ...current }
-      for (const section of sections) next[section] = 'idle'
-      return next
-    })
-    setTestErrors((current) => {
-      const next = { ...current }
-      for (const section of sections) delete next[section]
-      return next
-    })
-  }
-
-  useEffect(() => {
-    window.rex
-      .getSettings('integrations')
-      .then(async (settings: Settings) => {
-        const rawAccounts = settings.emailAccounts
-        const emailAccounts: EmailAccount[] = Array.isArray(rawAccounts)
-          ? (rawAccounts as EmailAccount[])
-              .filter((a) => typeof a === 'object' && a !== null && typeof a.id === 'string')
-              .map((a) => ({
-                id: a.id,
-                backend: a.backend ?? (a as unknown as { provider?: string }).provider ?? 'gmail',
-                displayName: a.displayName ?? '',
-                clientId: a.clientId ?? '',
-                clientSecret: a.clientSecret ?? '',
-                host: a.host ?? '',
-                port: typeof a.port === 'number' ? a.port : 993,
-                username: a.username ?? '',
-                password: a.password ?? '',
-                credentialRef: a.credentialRef,
-                hasCredential: a.hasCredential === true,
-                lastSynced: a.lastSynced
-              } as EmailAccount))
-          : []
-        setForm({
-          emailProvider:
-            settings.emailProvider === 'outlook' ? 'outlook' : 'gmail',
-          emailClientId: typeof settings.emailClientId === 'string' ? settings.emailClientId : '',
-          emailClientSecret:
-            typeof settings.emailClientSecret === 'string' ? settings.emailClientSecret : '', // pragma: allowlist secret
-          emailAccounts,
-          calendarProvider:
-            settings.calendarProvider === 'outlook' ? 'outlook' : 'gmail',
-          calendarClientId:
-            typeof settings.calendarClientId === 'string' ? settings.calendarClientId : '',
-          calendarClientSecret:
-            typeof settings.calendarClientSecret === 'string' ? settings.calendarClientSecret : '', // pragma: allowlist secret
-          smsSid: typeof settings.smsSid === 'string' ? settings.smsSid : '',
-          smsAuthToken: typeof settings.smsAuthToken === 'string' ? settings.smsAuthToken : '',
-          smsFromNumber: typeof settings.smsFromNumber === 'string' ? settings.smsFromNumber : '',
-          haUrl: typeof settings.haUrl === 'string' ? settings.haUrl : '',
-          haToken: typeof settings.haToken === 'string' ? settings.haToken : '',
-          phoneSid: typeof settings.phoneSid === 'string' ? settings.phoneSid : '',
-          phoneAuthToken: typeof settings.phoneAuthToken === 'string' ? settings.phoneAuthToken : '',
-          phoneNumber: typeof settings.phoneNumber === 'string' ? settings.phoneNumber : '',
-          phoneTransferNumber: typeof settings.phoneTransferNumber === 'string' ? settings.phoneTransferNumber : '',
-          voicemailNotificationsEnabled: typeof settings.voicemailNotificationsEnabled === 'boolean' ? settings.voicemailNotificationsEnabled : false,
-          contactsFilePath: typeof settings.contactsFilePath === 'string' ? settings.contactsFilePath : '',
-          telegramBotToken: typeof settings.telegramBotToken === 'string' ? settings.telegramBotToken : '',
-          telegramChatId: typeof settings.telegramChatId === 'string' ? settings.telegramChatId : '',
-          openclawGatewayUrl: typeof settings.openclawGatewayUrl === 'string' ? settings.openclawGatewayUrl : '',
-          openclawToolsEnabled: settings.openclawToolsEnabled === true,
-          openclawVoiceEnabled: settings.openclawVoiceEnabled === true,
-          openclawToken: typeof settings.openclawToken === 'string' ? settings.openclawToken : '',
-          credentialStatus:
-            settings.credentialStatus && typeof settings.credentialStatus === 'object'
-              ? settings.credentialStatus as IntegrationsSettings['credentialStatus']
-              : {}
-        })
-
-        const inventory = await window.rex.getIntegrations().catch(() => null)
-        if (inventory?.ok) {
-          applyIntegrationStatuses(inventory.integrations)
-        }
-      })
-      .catch(() => {
-        addToast('Failed to load integrations settings', 'error')
-      })
-      .finally(() => setLoading(false))
-  }, [addToast])
-
-  function showSaved(field: keyof IntegrationsSettings): void {
-    if (savedTimerRef.current) clearTimeout(savedTimerRef.current)
-    setSavedField(field)
-    savedTimerRef.current = setTimeout(() => setSavedField(null), 2000)
-  }
-
-  function saveField(field: keyof IntegrationsSettings, updatedForm: IntegrationsSettings): void {
-    window.rex
-      .setSettings('integrations', updatedForm as unknown as Settings)
-      .then((result) => {
-        if (result.ok) {
-          showSaved(field)
-          if (isIntegrationSecretField(field)) {
-            setForm((current) => ({
-              ...current,
-              [field]: '',
-              credentialStatus: {
-                ...current.credentialStatus,
-                [field]: { ref: current.credentialStatus[field]?.ref ?? '', hasCredential: true }
-              }
-            }))
-          }
-        } else {
-          addToast(result.error ?? 'Failed to save integrations settings', 'error')
-        }
-      })
-      .catch(() => {
-        addToast('Failed to save integrations settings', 'error')
-      })
-  }
-
-  function handleFieldChange<K extends keyof IntegrationsSettings>(
-    field: K,
-    value: IntegrationsSettings[K]
-  ): void {
-    const updated = { ...form, [field]: value }
-    setForm(updated)
-    resetStatusForField(field)
-    saveField(field, updated)
-  }
-
-  function handleTest(section: IntegrationSection): void {
-    setTestStatus((s) => ({ ...s, [section]: 'testing' }))
-    setTestErrors((s) => {
-      const next = { ...s }
-      delete next[section]
-      return next
-    })
-    const testRequest =
-      section === 'homeassistant'
-        ? window.rex.testHomeAssistant(form.haUrl, form.haToken)
-        : window.rex.testIntegration(section)
-
-    testRequest
-      .then((res) => {
-        const evidenceState = 'state' in res
-          ? res.state as IntegrationInventoryItem['state'] | undefined
-          : undefined
-        const nextStatus = evidenceState
-          ? integrationStatusToTestStatus(evidenceState)
-          : res.ok ? 'ok' : 'error'
-        setTestStatus((s) => ({ ...s, [section]: nextStatus }))
-        setTestErrors((s) => {
-          const next = { ...s }
-          if ((!res.ok || evidenceState === 'configured') && res.error) {
-            next[section] = res.error
-          } else {
-            delete next[section]
-          }
-          return next
-        })
-      })
-      .catch((err) => {
-        setTestStatus((s) => ({ ...s, [section]: 'error' }))
-        setTestErrors((s) => ({
-          ...s,
-          [section]: err instanceof Error ? err.message : 'Connection test failed'
-        }))
-      })
-  }
-
-  function handleAddEmailAccount(): void {
-    const newAccount: EmailAccount = {
-      id: `${Date.now()}`,
-      backend: 'gmail',
-      displayName: '',
-      clientId: '',
-      clientSecret: '',
-      host: '',
-      port: 993,
-      username: '',
-      password: ''
-    }
-    const updated = { ...form, emailAccounts: [...form.emailAccounts, newAccount] }
-    setForm(updated)
-    resetStatusForField('emailAccounts')
-    window.rex
-      .setSettings('integrations', updated as unknown as Settings)
-      .then((result) => {
-        if (!result.ok) {
-          setForm(form)
-          addToast(result.error ?? 'Failed to save email account', 'error')
-        }
-      })
-      .catch(() => {
-        setForm(form)
-        addToast('Failed to save email account', 'error')
-      })
-  }
-
-  function updateEmailAccountDraft(id: string, patch: Partial<EmailAccount>): void {
-    setForm((current) => ({
-      ...current,
-      emailAccounts: current.emailAccounts.map((account) =>
-        account.id === id ? { ...account, ...patch } : account
-      )
-    }))
-    resetStatusForField('emailAccounts')
-  }
-
-  function handleUpdateEmailAccount(id: string, patch: Partial<EmailAccount>): void {
-    const updated = {
-      ...form,
-      emailAccounts: form.emailAccounts.map((a) => (a.id === id ? { ...a, ...patch } : a))
-    }
-    setForm(updated)
-    resetStatusForField('emailAccounts')
-    window.rex
-      .removeEmailAccount(id, true)
-      .then((result) => {
-        if (!result.ok) {
-          setForm(form)
-          addToast(result.error ?? 'Failed to save email account', 'error')
-        } else if (patch.password || patch.clientSecret) {
-          setForm((current) => ({
-            ...current,
-            emailAccounts: current.emailAccounts.map((account) =>
-              account.id === id
-                ? { ...account, password: '', clientSecret: '', hasCredential: true }
-                : account
-            )
-          }))
-        }
-      })
-      .catch(() => {
-        setForm(form)
-        addToast('Failed to save email account', 'error')
-      })
-  }
-
-  function handleRemoveEmailAccount(id: string): void {
-    if (!window.confirm('Remove this email account and delete its stored credential?')) return
-    const updated = {
-      ...form,
-      emailAccounts: form.emailAccounts.filter((a) => a.id !== id)
-    }
-    setForm(updated)
-    resetStatusForField('emailAccounts')
-    window.rex
-      .setSettings('integrations', updated as unknown as Settings)
-      .then((result) => {
-        if (!result.ok) {
-          setForm(form)
-          addToast(result.error ?? 'Failed to remove email account', 'error')
-        }
-      })
-      .catch(() => {
-        setForm(form)
-        addToast('Failed to remove email account', 'error')
-      })
-  }
-
-  const inputClass =
-    'w-full bg-surface-raised border border-border rounded-lg px-3 py-2 text-sm text-text-primary placeholder:text-text-secondary focus:outline-none focus:ring-2 focus:ring-accent'
-
-  if (loading) {
-    return <PageLoadingFallback lines={6} />
-  }
-
-  return (
-    <div className="p-6 max-w-lg">
-      <h2 className="text-lg font-semibold text-text-primary mb-6">Integrations</h2>
-
-      {/* Email section */}
-      <section className="mb-7">
-        <div className="mb-4 flex items-center justify-between gap-3">
-          <h3 className="text-sm font-semibold text-text-primary flex items-center gap-2">
-            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-              <path d="M4 4h16c1.1 0 2 .9 2 2v12c0 1.1-.9 2-2 2H4c-1.1 0-2-.9-2-2V6c0-1.1.9-2 2-2z" />
-              <polyline points="22,6 12,13 2,6" />
-            </svg>
-            Email
-          </h3>
-          <div className="flex items-center gap-2">
-            <a
-              href="https://console.cloud.google.com/apis/credentials"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-xs text-accent hover:underline"
-            >
-              Google Console →
-            </a>
-            <ConnectionBadge
-              status={testStatus.email}
-              hasCredentials={
-                (form.emailClientId.trim() !== '' && hasStoredCredential(form, 'emailClientSecret')) ||
-                form.emailAccounts.some((account) => account.hasCredential === true)
-              }
-            />
-          </div>
-        </div>
-
-        <p className="text-xs text-text-secondary mb-4">
-          Optional. Lets Rex read and send mail. Create OAuth credentials in the Google Console linked above.
-        </p>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <label htmlFor="emailProvider" className="text-sm font-medium text-text-primary">Provider</label>
-            <SavedIndicator visible={savedField === 'emailProvider'} />
-          </div>
-          <select
-            id="emailProvider"
-            value={form.emailProvider}
-            onChange={(e) => handleFieldChange('emailProvider', e.target.value as IntegrationsSettings['emailProvider'])}
-            className={inputClass}
-          >
-            <option value="gmail">Gmail</option>
-            <option value="outlook">Outlook</option>
-          </select>
-          {form.emailProvider === 'outlook' && (
-            <p className="mt-2 text-xs text-text-secondary">
-              Outlook mailbox sync is not live yet. These fields store app credentials only.
-            </p>
-          )}
-        </div>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <div className="flex items-center gap-1.5">
-              <label htmlFor="emailClientId" className="text-sm font-medium text-text-primary">OAuth Client ID</label>
-              <span
-                title="Create OAuth 2.0 credentials in Google Cloud Console → APIs & Services → Credentials. Enable the Gmail API first."
-                className="flex-shrink-0 w-4 h-4 rounded-full bg-surface-raised text-text-muted flex items-center justify-center text-[10px] font-bold cursor-help select-none"
-                aria-label="OAuth Client ID help"
-              >
-                ?
-              </span>
-            </div>
-            <SavedIndicator visible={savedField === 'emailClientId'} />
-          </div>
-          <input
-            id="emailClientId"
-            type="text"
-            value={form.emailClientId}
-            placeholder="Enter client ID"
-            onChange={(e) => setForm((f) => ({ ...f, emailClientId: e.target.value }))}
-            onBlur={(e) => handleFieldChange('emailClientId', e.target.value)}
-            className={inputClass}
-          />
-        </div>
-
-        <div className="mb-2">
-          <div className="flex items-center justify-between mb-1.5">
-            <label htmlFor="emailClientSecret" className="text-sm font-medium text-text-primary">OAuth Client Secret</label>
-            <SavedIndicator visible={savedField === 'emailClientSecret'} />
-          </div>
-          <PasswordInput
-            id="emailClientSecret"
-            value={form.emailClientSecret}
-            placeholder="Enter client secret (example)" // pragma: allowlist secret
-            onChange={(v) => setForm((f) => ({ ...f, emailClientSecret: v }))}
-            onBlur={() => handleFieldChange('emailClientSecret', form.emailClientSecret)}
-          />
-        </div>
-
-        <TestConnectionButton status={testStatus.email} error={testErrors.email} onTest={() => handleTest('email')} />
-
-        {/* Multi-account email list */}
-        <div className="mt-5">
-          <div className="flex items-center justify-between mb-3">
-            <span className="text-sm font-medium text-text-primary">Email Accounts</span>
-            <button
-              type="button"
-              onClick={handleAddEmailAccount}
-              className="flex items-center gap-1.5 rounded-lg border border-border bg-surface-raised px-3 py-1.5 text-xs font-medium text-text-primary transition-colors hover:bg-border focus:outline-none focus:ring-2 focus:ring-accent"
-            >
-              <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5">
-                <line x1="12" y1="5" x2="12" y2="19" />
-                <line x1="5" y1="12" x2="19" y2="12" />
-              </svg>
-              Add Account
-            </button>
-          </div>
-          {form.emailAccounts.length === 0 ? (
-            <div className="rounded-xl border border-dashed border-border bg-surface-raised/40 px-4 py-4 text-sm text-text-secondary">
-              No additional accounts. Click "Add Account" to connect another inbox.
-            </div>
-          ) : (
-            <div className="space-y-3">
-              {form.emailAccounts.map((account) => {
-                const acctTestStatus = accountTestStatus[account.id] ?? 'idle'
-                return (
-                  <div key={account.id} className="rounded-xl border border-border bg-surface-raised p-4">
-                    {/* Header row: display name + remove */}
-                    <div className="mb-3 flex items-center justify-between gap-2">
-                      <input
-                        type="text"
-                        value={account.displayName}
-                        placeholder="Account label (e.g. Work Gmail)"
-                        onChange={(e) =>
-                          handleUpdateEmailAccount(account.id, { displayName: e.target.value })
-                        }
-                        className="flex-1 bg-bg border border-border rounded-lg px-3 py-1.5 text-sm text-text-primary focus:outline-none focus:ring-2 focus:ring-accent"
-                      />
-                      <button
-                        type="button"
-                        onClick={() => handleRemoveEmailAccount(account.id)}
-                        className="rounded-lg border border-danger/30 px-2.5 py-1.5 text-xs font-medium text-danger transition-colors hover:bg-danger/10 focus:outline-none"
-                      >
-                        Remove
-                      </button>
-                    </div>
-
-                    {/* Backend selector */}
-                    <div className="mb-2">
-                      <select
-                        value={account.backend}
-                        onChange={(e) =>
-                          handleUpdateEmailAccount(account.id, {
-                            backend: e.target.value as EmailAccount['backend']
-                          })
-                        }
-                        className="w-full bg-bg border border-border rounded-lg px-3 py-1.5 text-sm text-text-primary focus:outline-none focus:ring-2 focus:ring-accent"
-                      >
-                        <option value="gmail">Gmail OAuth</option>
-                        <option value="outlook">Outlook OAuth</option>
-                        <option value="imap">IMAP</option>
-                      </select>
-                    </div>
-
-                    {/* Credential fields */}
-                    {account.backend === 'imap' ? (
-                      <div className="space-y-2 mb-2">
-                        <input
-                          type="text"
-                          value={account.host}
-                          placeholder="IMAP host (e.g. imap.gmail.com)"
-                          onChange={(e) =>
-                            handleUpdateEmailAccount(account.id, { host: e.target.value })
-                          }
-                          className={inputClass}
-                        />
-                        <input
-                          type="number"
-                          value={account.port}
-                          placeholder="Port (993)"
-                          onChange={(e) =>
-                            handleUpdateEmailAccount(account.id, { port: parseInt(e.target.value, 10) || 993 })
-                          }
-                          className={inputClass}
-                        />
-                        <input
-                          type="text"
-                          value={account.username}
-                          placeholder="Username / email address"
-                          onChange={(e) =>
-                            handleUpdateEmailAccount(account.id, { username: e.target.value })
-                          }
-                          className={inputClass}
-                        />
-                        <PasswordInput
-                          id={`imap-pass-${account.id}`}
-                          value={account.password}
-                          placeholder={account.hasCredential ? 'Stored credential (enter to replace)' : 'Password or app password'}
-                          onChange={(v) => updateEmailAccountDraft(account.id, { password: v })}
-                          onBlur={() => {
-                            if (account.password) handleUpdateEmailAccount(account.id, { password: account.password })
-                          }}
-                        />
-                      </div>
-                    ) : (
-                      <div className="space-y-2 mb-2">
-                        <input
-                          type="text"
-                          value={account.clientId}
-                          placeholder="OAuth Client ID"
-                          onChange={(e) =>
-                            handleUpdateEmailAccount(account.id, { clientId: e.target.value })
-                          }
-                          className={inputClass}
-                        />
-                        <PasswordInput
-                          id={`email-secret-${account.id}`}
-                          value={account.clientSecret}
-                          placeholder={account.hasCredential ? 'Stored credential (enter to replace)' : 'OAuth Client Secret'}
-                          onChange={(v) => updateEmailAccountDraft(account.id, { clientSecret: v })}
-                          onBlur={() => {
-                            if (account.clientSecret) {
-                              handleUpdateEmailAccount(account.id, { clientSecret: account.clientSecret })
-                            }
-                          }}
-                        />
-                      </div>
-                    )}
-
-                    {/* Configuration evidence + last-synced */}
-                    <div className="flex items-center gap-3">
-                      <button
-                        type="button"
-                        onClick={() => handleTestEmailAccount(account.id)}
-                        disabled={acctTestStatus === 'testing'}
-                        className="flex items-center gap-1.5 rounded-lg border border-border bg-surface-raised px-3 py-1.5 text-xs font-medium text-text-primary transition-colors hover:bg-border focus:outline-none focus:ring-2 focus:ring-accent disabled:opacity-50"
-                      >
-                        {acctTestStatus === 'testing' ? 'Checking…' : 'Check Configuration'}
-                      </button>
-                      {acctTestStatus === 'ok' && (
-                        <span className="text-xs font-medium text-success">Authenticated</span>
-                      )}
-                      {acctTestStatus === 'configured' && (
-                        <span className="text-xs font-medium text-warning">Configured only</span>
-                      )}
-                      {acctTestStatus === 'error' && (
-                        <span className="text-xs font-medium text-danger">Failed</span>
-                      )}
-                      {account.lastSynced && (
-                        <span className="ml-auto text-xs text-text-secondary">
-                          Synced {new Date(account.lastSynced).toLocaleString()}
-                        </span>
-                      )}
-                    </div>
-                  </div>
-                )
-              })}
-            </div>
-          )}
-        </div>
-      </section>
-
-      <div className="border-t border-border mb-7" />
-
-      {/* Calendar section */}
-      <section className="mb-7">
-        <div className="mb-4 flex items-center justify-between gap-3">
-          <h3 className="text-sm font-semibold text-text-primary flex items-center gap-2">
-            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-              <rect x="3" y="4" width="18" height="18" rx="2" ry="2" />
-              <line x1="16" y1="2" x2="16" y2="6" />
-              <line x1="8" y1="2" x2="8" y2="6" />
-              <line x1="3" y1="10" x2="21" y2="10" />
-            </svg>
-            Calendar
-          </h3>
-          <div className="flex items-center gap-2">
-            <a
-              href="https://console.cloud.google.com/apis/credentials"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-xs text-accent hover:underline"
-            >
-              Google Console →
-            </a>
-            <ConnectionBadge
-              status={testStatus.calendar}
-              hasCredentials={form.calendarClientId.trim() !== '' && hasStoredCredential(form, 'calendarClientSecret')}
-            />
-          </div>
-        </div>
-
-        <p className="text-xs text-text-secondary mb-4">
-          Optional. Lets Rex read and create events on your calendar. Use the Google Console link above for OAuth credentials.
-        </p>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <label htmlFor="calendarProvider" className="text-sm font-medium text-text-primary">Provider</label>
-            <SavedIndicator visible={savedField === 'calendarProvider'} />
-          </div>
-          <select
-            id="calendarProvider"
-            value={form.calendarProvider}
-            onChange={(e) => handleFieldChange('calendarProvider', e.target.value as IntegrationsSettings['calendarProvider'])}
-            className={inputClass}
-          >
-            <option value="gmail">Google Calendar</option>
-            <option value="outlook">Outlook Calendar</option>
-          </select>
-          {form.calendarProvider === 'outlook' && (
-            <p className="mt-2 text-xs text-text-secondary">
-              Outlook calendar sync is not live yet. Rex cannot read or write Outlook events from these fields.
-            </p>
-          )}
-        </div>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <div className="flex items-center gap-1.5">
-              <label htmlFor="calendarClientId" className="text-sm font-medium text-text-primary">OAuth Client ID</label>
-              <span
-                title="Create OAuth 2.0 credentials in Google Cloud Console → APIs & Services → Credentials. Enable the Google Calendar API first."
-                className="flex-shrink-0 w-4 h-4 rounded-full bg-surface-raised text-text-muted flex items-center justify-center text-[10px] font-bold cursor-help select-none"
-                aria-label="OAuth Client ID help"
-              >
-                ?
-              </span>
-            </div>
-            <SavedIndicator visible={savedField === 'calendarClientId'} />
-          </div>
-          <input
-            id="calendarClientId"
-            type="text"
-            value={form.calendarClientId}
-            placeholder="Enter client ID"
-            onChange={(e) => setForm((f) => ({ ...f, calendarClientId: e.target.value }))}
-            onBlur={(e) => handleFieldChange('calendarClientId', e.target.value)}
-            className={inputClass}
-          />
-        </div>
-
-        <div className="mb-2">
-          <div className="flex items-center justify-between mb-1.5">
-            <label htmlFor="calendarClientSecret" className="text-sm font-medium text-text-primary">OAuth Client Secret</label>
-            <SavedIndicator visible={savedField === 'calendarClientSecret'} />
-          </div>
-          <PasswordInput
-            id="calendarClientSecret"
-            value={form.calendarClientSecret}
-            placeholder="Enter client secret (example)" // pragma: allowlist secret
-            onChange={(v) => setForm((f) => ({ ...f, calendarClientSecret: v }))}
-            onBlur={() => handleFieldChange('calendarClientSecret', form.calendarClientSecret)}
-          />
-        </div>
-
-        <TestConnectionButton status={testStatus.calendar} error={testErrors.calendar} onTest={() => handleTest('calendar')} />
-      </section>
-
-      <div className="border-t border-border mb-7" />
-
-      {/* SMS section */}
-      <section className="mb-7">
-        <div className="mb-4 flex items-center justify-between gap-3">
-          <h3 className="text-sm font-semibold text-text-primary flex items-center gap-2">
-            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-              <path d="M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z" />
-            </svg>
-            SMS (Twilio)
-          </h3>
-          <div className="flex items-center gap-2">
-            <a
-              href="https://console.twilio.com"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-xs text-accent hover:underline"
-            >
-              Twilio Console →
-            </a>
-            <ConnectionBadge
-              status={testStatus.sms}
-              hasCredentials={hasStoredCredential(form, 'smsSid') && hasStoredCredential(form, 'smsAuthToken') && hasStoredCredential(form, 'smsFromNumber')}
-            />
-          </div>
-        </div>
-
-        <p className="text-xs text-text-secondary mb-4">
-          Optional. Lets Rex send text messages through Twilio. Grab the SID, token, and a sender number from the Twilio Console link above.
-        </p>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <div className="flex items-center gap-1.5">
-              <label htmlFor="smsSid" className="text-sm font-medium text-text-primary">Account SID</label>
-              <span
-                title="Find your Account SID in the Twilio Console dashboard under 'Account Info'."
-                className="flex-shrink-0 w-4 h-4 rounded-full bg-surface-raised text-text-muted flex items-center justify-center text-[10px] font-bold cursor-help select-none"
-                aria-label="Where to find Account SID"
-              >
-                ?
-              </span>
-            </div>
-            <SavedIndicator visible={savedField === 'smsSid'} />
-          </div>
-          <input
-            id="smsSid"
-            type="text"
-            value={form.smsSid}
-            placeholder="Example Account SID" // pragma: allowlist secret
-            onChange={(e) => setForm((f) => ({ ...f, smsSid: e.target.value }))}
-            onBlur={(e) => handleFieldChange('smsSid', e.target.value)}
-            className={inputClass}
-          />
-        </div>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <label htmlFor="smsAuthToken" className="text-sm font-medium text-text-primary">Auth Token</label>
-            <SavedIndicator visible={savedField === 'smsAuthToken'} />
-          </div>
-          <PasswordInput
-            id="smsAuthToken"
-            value={form.smsAuthToken}
-            placeholder="Enter auth token (example)" // pragma: allowlist secret
-            onChange={(v) => setForm((f) => ({ ...f, smsAuthToken: v }))}
-            onBlur={() => handleFieldChange('smsAuthToken', form.smsAuthToken)}
-          />
-        </div>
-
-        <div className="mb-2">
-          <div className="flex items-center justify-between mb-1.5">
-            <label htmlFor="smsFromNumber" className="text-sm font-medium text-text-primary">From Phone Number</label>
-            <SavedIndicator visible={savedField === 'smsFromNumber'} />
-          </div>
-          <input
-            id="smsFromNumber"
-            type="text"
-            value={form.smsFromNumber}
-            placeholder="+15551234567"
-            onChange={(e) => setForm((f) => ({ ...f, smsFromNumber: e.target.value }))}
-            onBlur={(e) => handleFieldChange('smsFromNumber', e.target.value)}
-            className={inputClass}
-          />
-        </div>
-
-        <TestConnectionButton status={testStatus.sms} error={testErrors.sms} onTest={() => handleTest('sms')} />
-      </section>
-
-      <div className="border-t border-border mb-7" />
-
-      {/* Home Assistant section */}
-      <section className="mb-2">
-        <div className="mb-4 flex items-center justify-between gap-3">
-          <h3 className="text-sm font-semibold text-text-primary flex items-center gap-2">
-            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-              <path d="M3 9l9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z" />
-              <polyline points="9 22 9 12 15 12 15 22" />
-            </svg>
-            Home Assistant
-          </h3>
-          <div className="flex items-center gap-2">
-            <NavLink
-              to="/settings/home-assistant"
-              className="text-xs text-accent hover:underline"
-            >
-              Full setup →
-            </NavLink>
-            <ConnectionBadge
-              status={testStatus.homeassistant}
-              hasCredentials={form.haUrl.trim() !== '' && hasStoredCredential(form, 'haToken')}
-            />
-          </div>
-        </div>
-
-        <p className="text-xs text-text-secondary mb-4">
-          Optional. Lets Rex control your smart home. Create a long-lived access token from your Home Assistant user profile page.
-        </p>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <label htmlFor="haUrl" className="text-sm font-medium text-text-primary">Base URL</label>
-            <SavedIndicator visible={savedField === 'haUrl'} />
-          </div>
-          <input
-            id="haUrl"
-            type="text"
-            value={form.haUrl}
-            placeholder="http://homeassistant.local:8123"
-            onChange={(e) => {
-              setForm((f) => ({ ...f, haUrl: e.target.value }))
-              setTestStatus((s) => ({ ...s, homeassistant: 'idle' }))
-            }}
-            onBlur={(e) => handleFieldChange('haUrl', e.target.value)}
-            className={inputClass}
-          />
-        </div>
-
-        <div className="mb-2">
-          <div className="flex items-center justify-between mb-1.5">
-            <label htmlFor="haToken" className="text-sm font-medium text-text-primary">Long-Lived Access Token</label>
-            <SavedIndicator visible={savedField === 'haToken'} />
-          </div>
-          <PasswordInput
-            id="haToken"
-            value={form.haToken}
-            placeholder="Enter access token" // pragma: allowlist secret
-            onChange={(v) => {
-              setForm((f) => ({ ...f, haToken: v }))
-              setTestStatus((s) => ({ ...s, homeassistant: 'idle' }))
-            }}
-            onBlur={() => handleFieldChange('haToken', form.haToken)}
-          />
-        </div>
-
-        <TestConnectionButton status={testStatus.homeassistant} error={testErrors.homeassistant} onTest={() => handleTest('homeassistant')} />
-      </section>
-
-      <div className="border-t border-border mb-7" />
-
-      {/* OpenClaw section */}
-      <section className="mb-7">
-        <div className="mb-4 flex items-center justify-between gap-3">
-          <div>
-            <h3 className="text-sm font-semibold text-text-primary">OpenClaw</h3>
-            <div className="mt-1 text-xs font-medium text-warning">Experimental - off by default</div>
-          </div>
-          <ConnectionBadge
-            status={testStatus.openclaw}
-            hasCredentials={form.openclawGatewayUrl.trim() !== '' && hasStoredCredential(form, 'openclawToken')}
-          />
-        </div>
-        <p className="mb-4 text-xs text-text-secondary">
-          Optional external capability gateway. The token stays in the credential vault and is never returned to the renderer after saving.
-        </p>
-        <div className="mb-4">
-          <div className="mb-1.5 flex items-center justify-between">
-            <label htmlFor="openclawGatewayUrl" className="text-sm font-medium text-text-primary">Gateway URL</label>
-            <SavedIndicator visible={savedField === 'openclawGatewayUrl'} />
-          </div>
-          <input
-            id="openclawGatewayUrl"
-            type="url"
-            value={form.openclawGatewayUrl}
-            placeholder="http://127.0.0.1:18789"
-            onChange={(e) => setForm((f) => ({ ...f, openclawGatewayUrl: e.target.value }))}
-            onBlur={(e) => handleFieldChange('openclawGatewayUrl', e.target.value)}
-            className={inputClass}
-          />
-        </div>
-        <div className="mb-4">
-          <div className="mb-1.5 flex items-center justify-between">
-            <label htmlFor="openclawToken" className="text-sm font-medium text-text-primary">Gateway Token</label>
-            <SavedIndicator visible={savedField === 'openclawToken'} />
-          </div>
-          <PasswordInput
-            id="openclawToken"
-            value={form.openclawToken}
-            placeholder={hasStoredCredential(form, 'openclawToken') ? 'Stored credential (enter to replace)' : 'Enter gateway token'}
-            onChange={(v) => setForm((f) => ({ ...f, openclawToken: v }))}
-            onBlur={() => { if (form.openclawToken) handleFieldChange('openclawToken', form.openclawToken) }}
-          />
-        </div>
-        <div className="mb-4 space-y-2">
-          <label className="flex items-center gap-2 text-sm text-text-primary">
-            <input
-              type="checkbox"
-              checked={form.openclawToolsEnabled}
-              disabled={form.openclawGatewayUrl.trim() === '' || (!hasStoredCredential(form, 'openclawToken') && form.openclawToken.trim() === '')}
-              onChange={(e) => handleFieldChange('openclawToolsEnabled', e.target.checked)}
-            />
-            Enable OpenClaw tools
-          </label>
-          <label className="flex items-center gap-2 text-sm text-text-primary">
-            <input
-              type="checkbox"
-              checked={form.openclawVoiceEnabled}
-              disabled={form.openclawGatewayUrl.trim() === '' || (!hasStoredCredential(form, 'openclawToken') && form.openclawToken.trim() === '')}
-              onChange={(e) => handleFieldChange('openclawVoiceEnabled', e.target.checked)}
-            />
-            Enable OpenClaw voice backend
-          </label>
-        </div>
-        <TestConnectionButton
-          status={testStatus.openclaw}
-          error={testErrors.openclaw}
-          onTest={() => handleTest('openclaw')}
-        />
-        {testStatus.openclaw === 'configured' && !testErrors.openclaw && (
-          <p className="mt-2 text-xs text-text-secondary">
-            Gateway reachable; authentication and tool capability are not yet proven.
-          </p>
-        )}
-      </section>
-
-      {/* Phone (Twilio) section */}
-      <section className="mb-2">
-        <div className="mb-4 flex items-center justify-between gap-3">
-          <h3 className="text-sm font-semibold text-text-primary flex items-center gap-2">
-            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-              <path d="M22 16.92v3a2 2 0 0 1-2.18 2 19.79 19.79 0 0 1-8.63-3.07A19.5 19.5 0 0 1 4.69 13.6 19.79 19.79 0 0 1 1.61 5.08 2 2 0 0 1 3.58 3h3a2 2 0 0 1 2 1.72c.127.96.361 1.903.7 2.81a2 2 0 0 1-.45 2.11L7.91 10.09a16 16 0 0 0 6 6l.91-.91a2 2 0 0 1 2.11-.45c.907.339 1.85.573 2.81.7A2 2 0 0 1 22 16.92z" />
-            </svg>
-            Phone (Twilio)
-          </h3>
-          <div className="flex items-center gap-2">
-            <a
-              href="https://console.twilio.com"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-xs text-accent hover:underline"
-            >
-              Twilio Console →
-            </a>
-            <ConnectionBadge
-              status={testStatus.phone}
-              hasCredentials={hasStoredCredential(form, 'phoneSid') && hasStoredCredential(form, 'phoneAuthToken') && hasStoredCredential(form, 'phoneNumber')}
-            />
-          </div>
-        </div>
-
-        <p className="text-xs text-text-secondary mb-4">
-          Optional. Lets Rex place voice calls through Twilio. Use the same Twilio Console link above to copy your credentials.
-        </p>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <div className="flex items-center gap-1.5">
-              <label htmlFor="phoneSid" className="text-sm font-medium text-text-primary">Account SID</label>
-              <span
-                title="Find your Account SID in the Twilio Console dashboard under 'Account Info'."
-                className="flex-shrink-0 w-4 h-4 rounded-full bg-surface-raised text-text-muted flex items-center justify-center text-[10px] font-bold cursor-help select-none"
-                aria-label="Where to find Account SID"
-              >
-                ?
-              </span>
-            </div>
-            <SavedIndicator visible={savedField === 'phoneSid'} />
-          </div>
-          <input
-            id="phoneSid"
-            type="text"
-            value={form.phoneSid}
-            placeholder="Example Account SID" // pragma: allowlist secret
-            onChange={(e) => setForm((f) => ({ ...f, phoneSid: e.target.value }))}
-            onBlur={(e) => handleFieldChange('phoneSid', e.target.value)}
-            className={inputClass}
-          />
-        </div>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <label htmlFor="phoneAuthToken" className="text-sm font-medium text-text-primary">Auth Token</label>
-            <SavedIndicator visible={savedField === 'phoneAuthToken'} />
-          </div>
-          <PasswordInput
-            id="phoneAuthToken"
-            value={form.phoneAuthToken}
-            placeholder="Enter auth token (example)" // pragma: allowlist secret
-            onChange={(v) => setForm((f) => ({ ...f, phoneAuthToken: v }))}
-            onBlur={() => handleFieldChange('phoneAuthToken', form.phoneAuthToken)}
-          />
-        </div>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <label htmlFor="phoneNumber" className="text-sm font-medium text-text-primary">Twilio Phone Number</label>
-            <SavedIndicator visible={savedField === 'phoneNumber'} />
-          </div>
-          <input
-            id="phoneNumber"
-            type="text"
-            value={form.phoneNumber}
-            placeholder="+15551234567"
-            onChange={(e) => setForm((f) => ({ ...f, phoneNumber: e.target.value }))}
-            onBlur={(e) => handleFieldChange('phoneNumber', e.target.value)}
-            className={inputClass}
-          />
-        </div>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <label htmlFor="phoneTransferNumber" className="text-sm font-medium text-text-primary">Transfer-to Number (optional)</label>
-            <SavedIndicator visible={savedField === 'phoneTransferNumber'} />
-          </div>
-          <input
-            id="phoneTransferNumber"
-            type="text"
-            value={form.phoneTransferNumber}
-            placeholder="+15559876543"
-            onChange={(e) => setForm((f) => ({ ...f, phoneTransferNumber: e.target.value }))}
-            onBlur={(e) => handleFieldChange('phoneTransferNumber', e.target.value)}
-            className={inputClass}
-          />
-        </div>
-
-        <div className="mb-4 flex items-center justify-between">
-          <label htmlFor="voicemailNotificationsEnabled" className="text-sm font-medium text-text-primary">Voicemail notifications</label>
-          <div className="flex items-center gap-2">
-            <SavedIndicator visible={savedField === 'voicemailNotificationsEnabled'} />
-            <input
-              id="voicemailNotificationsEnabled"
-              type="checkbox"
-              checked={form.voicemailNotificationsEnabled}
-              onChange={(e) => handleFieldChange('voicemailNotificationsEnabled', e.target.checked)}
-              className="h-4 w-4 rounded border-border text-accent accent-accent cursor-pointer"
-            />
-          </div>
-        </div>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <label className="text-sm font-medium text-text-primary">Contacts File</label>
-            <SavedIndicator visible={savedField === 'contactsFilePath'} />
-          </div>
-          <div className="flex items-center gap-2">
-            <input
-              type="text"
-              value={form.contactsFilePath}
-              readOnly
-              placeholder="No file selected"
-              className={`${inputClass} flex-1 cursor-default`}
-            />
-            <button
-              type="button"
-              onClick={() => {
-                window.rex.uploadContactsFile().then((res) => {
-                  if (res.ok && res.path) {
-                    handleFieldChange('contactsFilePath', res.path)
-                  }
-                })
-              }}
-              className="shrink-0 rounded-md border border-border bg-surface px-3 py-1.5 text-xs font-medium text-text-primary hover:bg-hover transition-colors"
-            >
-              Browse…
-            </button>
-          </div>
-          <p className="mt-1 text-xs text-text-secondary">Accepts .json or .vcf (vCard) contact files for outbound calling.</p>
-        </div>
-
-        <TestConnectionButton status={testStatus.phone} error={testErrors.phone} onTest={() => handleTest('phone')} />
-      </section>
-
-      <div className="border-t border-border mb-7" />
-
-      {/* Telegram section */}
-      <section className="mb-2">
-        <div className="mb-4 flex items-center justify-between gap-3">
-          <h3 className="text-sm font-semibold text-text-primary flex items-center gap-2">
-            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-              <path d="M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z" />
-            </svg>
-            Telegram
-          </h3>
-          <ConnectionBadge
-            status="idle"
-            hasCredentials={hasStoredCredential(form, 'telegramBotToken') && form.telegramChatId.trim() !== ''}
-          />
-        </div>
-
-        <p className="text-xs text-text-secondary mb-4">
-          Optional. Lets Rex send messages to a Telegram chat. Create a bot through @BotFather to get a token, then message it to find your chat ID.
-        </p>
-
-        <div className="mb-4">
-          <div className="flex items-center justify-between mb-1.5">
-            <div className="flex items-center gap-1.5">
-              <label htmlFor="telegramBotToken" className="text-sm font-medium text-text-primary">Bot Token</label>
-              <span
-              title="Create a bot via @BotFather on Telegram to get a token. Rex stores it in the Windows credential vault."
-                className="flex-shrink-0 w-4 h-4 rounded-full bg-surface-raised text-text-muted flex items-center justify-center text-[10px] font-bold cursor-help select-none"
-                aria-label="Where to find Bot Token"
-              >
-                ?
-              </span>
-            </div>
-            <SavedIndicator visible={savedField === 'telegramBotToken'} />
-          </div>
-          <PasswordInput
-            id="telegramBotToken"
-            value={form.telegramBotToken}
-            placeholder="Enter access token (example)" // pragma: allowlist secret
-            onChange={(v) => setForm((f) => ({ ...f, telegramBotToken: v }))}
-            onBlur={() => handleFieldChange('telegramBotToken', form.telegramBotToken)}
-          />
-        </div>
-
-        <div className="mb-2">
-          <div className="flex items-center justify-between mb-1.5">
-            <div className="flex items-center gap-1.5">
-              <label htmlFor="telegramChatId" className="text-sm font-medium text-text-primary">Chat ID</label>
-              <span
-                title="Your personal chat ID or group chat ID. Send a message to your bot and check the Telegram Bot API getUpdates response."
-                className="flex-shrink-0 w-4 h-4 rounded-full bg-surface-raised text-text-muted flex items-center justify-center text-[10px] font-bold cursor-help select-none"
-                aria-label="Where to find Chat ID"
-              >
-                ?
-              </span>
-            </div>
-            <SavedIndicator visible={savedField === 'telegramChatId'} />
-          </div>
-          <input
-            id="telegramChatId"
-            type="text"
-            value={form.telegramChatId}
-            placeholder="-1001234567890"
-            onChange={(e) => setForm((f) => ({ ...f, telegramChatId: e.target.value }))}
-            onBlur={(e) => handleFieldChange('telegramChatId', e.target.value)}
-            className={inputClass}
-          />
-          <p className="mt-1 text-xs text-text-secondary">
-            The token is stored in the Windows credential vault and is never loaded back into this field.
-          </p>
-        </div>
-      </section>
-    </div>
-  )
+  return <IntegrationsSettingsSection />
 }
 
 function NotificationsPanel(): React.ReactElement {
