+++ gui/src/pages/settings/shared.tsx
@@ -1,0 +1,102 @@
+import React, { useState } from 'react'
+
+export function Toggle({
+  checked,
+  onChange,
+  id
+}: {
+  checked: boolean
+  onChange: (v: boolean) => void
+  id: string
+}): React.ReactElement {
+  return (
+    <button
+      id={id}
+      role="switch"
+      aria-checked={checked}
+      onClick={() => onChange(!checked)}
+      className={[
+        'relative inline-flex h-6 w-11 shrink-0 items-center rounded-full transition-colors focus:outline-none focus:ring-2 focus:ring-accent focus:ring-offset-2 focus:ring-offset-bg',
+        checked ? 'bg-accent' : 'bg-surface-raised border border-border'
+      ].join(' ')}
+    >
+      <span
+        className={[
+          'inline-block h-4 w-4 rounded-full bg-white shadow transition-transform',
+          checked ? 'translate-x-6' : 'translate-x-1'
+        ].join(' ')}
+      />
+    </button>
+  )
+}
+
+export function SavedIndicator({ visible }: { visible: boolean }): React.ReactElement {
+  return (
+    <span
+      className={[
+        'flex items-center gap-1 text-xs text-success transition-opacity duration-300',
+        visible ? 'opacity-100' : 'opacity-0'
+      ].join(' ')}
+    >
+      <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3">
+        <polyline points="20 6 9 17 4 12" />
+      </svg>
+      Saved
+    </span>
+  )
+}
+
+function EyeIcon({ open }: { open: boolean }): React.ReactElement {
+  if (open) {
+    return (
+      <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
+        <path d="M1 12s4-8 11-8 11 8 11 8-4 8-11 8-11-8-11-8z" />
+        <circle cx="12" cy="12" r="3" />
+      </svg>
+    )
+  }
+  return (
+    <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
+      <path d="M17.94 17.94A10.07 10.07 0 0 1 12 20c-7 0-11-8-11-8a18.45 18.45 0 0 1 5.06-5.94" />
+      <path d="M9.9 4.24A9.12 9.12 0 0 1 12 4c7 0 11 8 11 8a18.5 18.5 0 0 1-2.16 3.19" />
+      <line x1="1" y1="1" x2="23" y2="23" />
+    </svg>
+  )
+}
+
+export function PasswordInput({
+  id,
+  value,
+  placeholder,
+  onChange,
+  onBlur
+}: {
+  id: string
+  value: string
+  placeholder?: string
+  onChange: (v: string) => void
+  onBlur?: () => void
+}): React.ReactElement {
+  const [show, setShow] = useState(false)
+  return (
+    <div className="relative">
+      <input
+        id={id}
+        type={show ? 'text' : 'password'}
+        value={value}
+        placeholder={placeholder}
+        onChange={(e) => onChange(e.target.value)}
+        onBlur={onBlur}
+        className="w-full bg-surface-raised border border-border rounded-lg px-3 py-2 pr-10 text-sm text-text-primary placeholder:text-text-secondary focus:outline-none focus:ring-2 focus:ring-accent"
+      />
+      <button
+        type="button"
+        onClick={() => setShow((s) => !s)}
+        className="absolute right-2.5 top-1/2 -translate-y-1/2 text-text-secondary hover:text-text-primary transition-colors"
+        aria-label={show ? 'Hide' : 'Show'}
+      >
+        <EyeIcon open={show} />
+      </button>
+    </div>
+  )
+}
